refactor: separate the IsAvailable method and related tests to separate files
