+++ paymaster/paymaster_test.go
@@ -3,7 +3,6 @@ package paymaster
 import (
 	"context"
 	"encoding/json"
-	"strconv"
 	"testing"
 
 	"github.com/NethermindEth/starknet.go/internal/tests"
@@ -11,36 +10,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 )
-
-// Test the 'paymaster_isAvailable' method
-func TestIsAvailable(t *testing.T) {
-	t.Parallel()
-	t.Run("integration", func(t *testing.T) {
-		tests.RunTestOn(t, tests.IntegrationEnv)
-		t.Parallel()
-
-		pm, spy := SetupPaymaster(t)
-		available, err := pm.IsAvailable(context.Background())
-		require.NoError(t, err)
-
-		assert.Equal(t, string(spy.LastResponse()), strconv.FormatBool(available))
-		assert.True(t, available)
-	})
-
-	t.Run("mock", func(t *testing.T) {
-		tests.RunTestOn(t, tests.MockEnv)
-		t.Parallel()
-
-		pm := SetupMockPaymaster(t)
-		pm.c.EXPECT().
-			CallContextWithSliceArgs(context.Background(), gomock.AssignableToTypeOf(new(bool)), "paymaster_isAvailable").
-			SetArg(1, true).
-			Return(nil)
-		available, err := pm.IsAvailable(context.Background())
-		assert.NoError(t, err)
-		assert.True(t, available)
-	})
-}
 
 // Test the 'paymaster_getSupportedTokens' method
 func TestGetSupportedTokens(t *testing.T) {
